requirements.txt
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
PyGithub>=2.1.0
google-generativeai>=0.3.0
pyyaml>=6.0
python-slugify>=8.0.0
